+++ changelog.md
@@ -3,6 +3,9 @@
 ## WIP
 - Make zero instance for `TransformMatrixKHR` return identity matrix. See
   https://github.com/expipiplus1/vulkan/issues/240
+- Remove explicit 'count' field in AccelerationStructureBuildGeometryInfoKHR.
+  See https://github.com/expipiplus1/vulkan/issues/239
+
 
 ## [3.8.3] - 2021-01-04
 - Bump API version to v1.2.166